Rename service_stats to ft_service_stats and update related references
