+++ app/dao/fact_service_stats_dao.py
@@ -6,7 +6,7 @@ from sqlalchemy import func
 from sqlalchemy.dialects.postgresql import insert
 
 from app import db
-from app.models import ServiceStats
+from app.models import FactServiceStats
 
 
 class ServiceStatsDimensions(TypedDict):
@@ -35,14 +35,14 @@ def _update_service_stats_count(dimensions: ServiceStatsDimensions, delta: int) 
         "notification_status": dimensions["notification_status"],
     }
     filters = (
-        ServiceStats.service_id == dimension_values["service_id"],
-        ServiceStats.template_id == dimension_values["template_id"],
-        ServiceStats.notification_type == dimension_values["notification_type"],
-        ServiceStats.notification_status == dimension_values["notification_status"],
+        FactServiceStats.service_id == dimension_values["service_id"],
+        FactServiceStats.template_id == dimension_values["template_id"],
+        FactServiceStats.notification_type == dimension_values["notification_type"],
+        FactServiceStats.notification_status == dimension_values["notification_status"],
     )
 
     if delta > 0:
-        stmt = insert(ServiceStats).values(
+        stmt = insert(FactServiceStats).values(
             id=uuid.uuid4(),
             **dimension_values,
             count=delta,
@@ -50,17 +50,17 @@ def _update_service_stats_count(dimensions: ServiceStatsDimensions, delta: int) 
         stmt = stmt.on_conflict_do_update(
             constraint="uix_ft_service_stats_dimensions",
             set_={
-                "count": ServiceStats.count + delta,
+                "count": FactServiceStats.count + delta,
             },
         )
         db.session.execute(stmt)
     else:
         (
-            db.session.query(ServiceStats)
+            db.session.query(FactServiceStats)
             .filter(*filters)
             .update(
                 {
-                    "count": func.greatest(ServiceStats.count + delta, 0),
+                    "count": func.greatest(FactServiceStats.count + delta, 0),
                 },
                 synchronize_session=False,
             )
@@ -69,7 +69,7 @@ def _update_service_stats_count(dimensions: ServiceStatsDimensions, delta: int) 
 
 # 3. Public read API that returns all stats rows for a single service, with a
 # defensive guard to avoid querying when no service id is provided.
-def dao_fetch_stats_for_service(service_id: UUID) -> list[ServiceStats]:
+def dao_fetch_stats_for_service(service_id: UUID) -> list[FactServiceStats]:
     """
     Fetch service stats for a specific service.
 
@@ -77,9 +77,9 @@ def dao_fetch_stats_for_service(service_id: UUID) -> list[ServiceStats]:
         service_id: UUID of the service to fetch stats for
 
     Returns:
-        List of ServiceStats records for the specified service
+        List of FactServiceStats records for the specified service
     """
     if not service_id:
         return []
 
-    return db.session.query(ServiceStats).filter(ServiceStats.service_id == service_id).all()
+    return db.session.query(FactServiceStats).filter(FactServiceStats.service_id == service_id).all()
